- Renamed function in ApplicationLayer to better reflect its purpose.
This was because the ApplicationLayer was holding too much power. The class should be able to tell the whole program that it should exit, which is what the name was implying. That kind of power should be left with whoever is handling the callback - which in this case, is main.